实现寄出类型的转换

+++ srcs/cpp/MagicRenderView/srcs/qt/varType/type/int64Type.h
@@ -5,7 +5,7 @@
 
 #include "../I_Stack.h"
 class Int64Type : public  I_Stack{
-	using t_current_type = int16_t;
+	using t_current_type = int64_t;
 public:
 	Int64Type( ) : I_Stack( typeid( t_current_type ) ) { }
 	bool createTarget( const type_info &target_type_info, const std_function<void(void *create_obj_ptr)> &create_call_function ) override;
